refactor: applied pr comments
